# Speed-up rotation page with redis