Before:
- We store the contents of each individual rotation file in redis. But, we give it a 1-hour expiration.

After:
- We still store the contents of each file in redis. But they don't have any expiration time.
- We **also** store the summary of the results - the number of runs and totaled hits for each card - in redis, so we don't even need to reiterate through the runs and tally things up again
- Rotation script needs to be aware of this, so it can invalidate keys as appropriate.

Before:
- With file contents in redis, 6.05s, 6.24s, 7.26s to load /rotation/ out of three attempts locally
- Cumulative time for `read_rotation_files` ~2.6s to 3s.

After:
- with redis fully loaded, 4.98s. 5.20s, 4.72s out of three attempts locally
- `read_rotation_files` doesn't need to be called.

What's next:
- The vast majority of the time left (3.6-4s locally) is spent in pystache.