@TravisBuddy 2019-11-04
Hey @tsutton,  
Your changes look good to me!

<a href="https:&#x2F;&#x2F;travis-ci.org&#x2F;PennyDreadfulMTG&#x2F;Penny-Dreadful-Tools&#x2F;builds&#x2F;606941032">View build log</a>

###### TravisBuddy Request Identifier: 0b5743e0-fea8-11e9-b5b5-d568244b8b3c